api/requirements.txt
Flask
gunicorn
google-cloud-bigtable
flask-cors